requirements.txt
python==3.8
scikit-learn==1.1.3
torch==1.12.1
torchvision==0.13.1
pillow==9.2.0
numpy==1.23.4
